Consolidate bitcoin scripts to single script
