--- packages/bitcoin-collateral-agent-provider/lib/index.js
@@ -1,3 +1,0 @@
-import BitcoinCollateralAgentProvider from './BitcoinCollateralAgentProvider'
-
-export default BitcoinCollateralAgentProvider
