--- packages/bitcoin-collateral-agent-provider/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@atomicloans/bitcoin-collateral-agent-provider",
-  "umdName": "BitcoinCollateralAgentProvider",
-  "umdExport": "default",
-  "version": "0.2.44",
-  "description": "",
-  "module": "lib/index.js",
-  "main": "dist/index.cjs.js",
-  "files": [
-    "dist",
-    "lib"
-  ],
-  "scripts": {
-    "build:node": "webpack --config ../../webpack/webpack.node.config.js",
-    "build:browser": "webpack --config ../../webpack/webpack.browser.config.js",
-    "build": "webpack --config ../../webpack/webpack.config.js"
-  },
-  "author": "AtomicLoans <info@atomicloans.io>",
-  "license": "MIT",
-  "dependencies": {
-    "@atomicloans/provider": "^0.2.43",
-    "@babel/runtime": "^7.4.3",
-    "@liquality/bitcoin-networks": "^0.2.11",
-    "@liquality/bitcoin-utils": "^0.2.11",
-    "@liquality/crypto": "^0.2.11",
-    "@liquality/provider": "^0.2.11",
-    "@liquality/utils": "^0.2.11"
-  },
-  "engines": {
-    "node": "~8.12.0"
-  },
-  "homepage": "https://github.com/liquality/chainabstractionlayer#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/liquality/chainabstractionlayer.git"
-  },
-  "bugs": {
-    "url": "https://github.com/liquality/chainabstractionlayer/issues"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "696c4e8c1f5e86724207817bff97d90711fa1f40"
-}
