+++ packages/bitcoin-collateral-provider/package.json
@@ -2,7 +2,7 @@
   "name": "@atomicloans/bitcoin-collateral-provider",
   "umdName": "BitcoinCollateralProvider",
   "umdExport": "default",
-  "version": "0.2.43",
+  "version": "0.2.44",
   "description": "",
   "module": "lib/index.js",
   "main": "dist/index.cjs.js",
