+++ packages/loan-bundle/lib/providers.js
@@ -1,10 +1,8 @@
 import BitcoinCollateralProvider from '@atomicloans/bitcoin-collateral-provider'
-import BitcoinCollateralAgentProvider from '@atomicloans/bitcoin-collateral-agent-provider'
 import BitcoinCollateralSwapProvider from '@atomicloans/bitcoin-collateral-swap-provider'
 
 const bitcoin = {
   BitcoinCollateralProvider,
-  BitcoinCollateralAgentProvider,
   BitcoinCollateralSwapProvider
 }
